+++ build-logic/convention/src/main/kotlin/EyesonConstants.kt
@@ -1,0 +1,8 @@
+object EyesonConstants {
+    const val VERSION = "1.2.0"
+    const val GROUP_ID = "com.eyeson"
+    const val DEMO_APP_VERSION = 10
+    const val COMPILE_SDK = 35
+    const val TARGET_SDK = 35
+    const val MIN_SDK = 23
+}