+++ build-logic/convention/src/main/kotlin/LibraryConventionPlugin.kt
@@ -1,0 +1,21 @@
+import com.android.build.gradle.LibraryExtension
+import extensions.configureKotlin
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
+
+class LibraryConventionPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            with(pluginManager) {
+                apply("com.android.library")
+                apply("org.jetbrains.kotlin.android")
+            }
+
+            extensions.configure<LibraryExtension> {
+                configureKotlin(this)
+                defaultConfig.targetSdk = EyesonConstants.TARGET_SDK
+            }
+        }
+    }
+}