+++ buildSrc/src/main/java/LibrariesKapt.kt
@@ -1,5 +1,0 @@
-object LibrariesKapt {
-    // Networking
-    const val moshiKotlinCodegen = "com.squareup.moshi:moshi-kotlin-codegen:${Versions.moshi}"
-    const val hiltAndroidCompiler = "com.google.dagger:hilt-android-compiler:${Versions.hilt}"
-}